feat, trubleshooting : 기능 생성중
기본 기능 생성중
Order : 관련 table name 으로 인한 문제 해결,
Order : service에 save, 기능 활성화,
Order : inqury 기능은 생성중,
아직 각 저장값에 유효성 검사는 구현 안됨

+++ src/main/java/com/example/week_v_homework/dto/order/OrderDetailRequestDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 public class OrderDetailRequestDto {
-    private int foodId;
+    private Long foodId;
     private int quantity;
 }
